+++ tests/testthat/test-rm_postal_code.R
@@ -1,0 +1,26 @@
+context("Checking rm_postal_code")
+
+test_that("rm_postal_code is removing postal_code strings",{
+
+    x <- c("Anchorage, AK", "New York City, NY", "Some Place, Another Place, LA")
+    rm_postal_code(x)
+    expect_identical(
+    	c("Anchorage,", "New York City,", "Some Place, Another Place,"), 
+    	rm_postal_code(x)
+    )
+	
+})
+
+
+test_that("rm_postal_code is extracting postal_code strings",{
+
+    x <- c("Anchorage, AK", "New York City, NY", "Some Place, Another Place, LA")
+    expect_identical(
+    	list("AK", "NY", "LA"), 
+    	rm_postal_code(x, extract=TRUE)
+    )
+
+})
+
+
+